new uint64 type

+++ src/kernel/Metadata.cpp
@@ -6,8 +6,8 @@ namespace cngn {
 Metadata::Metadata(const Schema& schema) : schema_(schema) {
 }
 
-Metadata::Metadata(Schema schema, std::vector<int64_t> batch_offsets,
-                   std::vector<int64_t> columns_cnt, std::vector<int64_t> rows_cnt)
+Metadata::Metadata(Schema schema, std::vector<uint64_t> batch_offsets,
+                   std::vector<uint64_t> columns_cnt, std::vector<uint64_t> rows_cnt)
     : schema_(std::move(schema)),
       batch_offsets_(std::move(batch_offsets)),
       columns_cnt_(std::move(columns_cnt)),
@@ -18,15 +18,15 @@ const Schema& Metadata::GetSchema() const {
     return schema_;
 }
 
-const std::vector<int64_t>& Metadata::GetOffsets() const {
+const std::vector<uint64_t>& Metadata::GetOffsets() const {
     return batch_offsets_;
 }
 
-const std::vector<int64_t>& Metadata::GetColumnsCnt() const {
+const std::vector<uint64_t>& Metadata::GetColumnsCnt() const {
     return columns_cnt_;
 }
 
-const std::vector<int64_t>& Metadata::GetRowsCnt() const {
+const std::vector<uint64_t>& Metadata::GetRowsCnt() const {
     return rows_cnt_;
 }
 
@@ -43,7 +43,7 @@ void Metadata::AddBatch(size_t offset, size_t columns, size_t rows) {
                << "Rows count: " << rows << '\n';
 }
 
-int64_t Metadata::GetNowOffset() const {
+uint64_t Metadata::GetNowOffset() const {
     return now_offset_;
 }
 
@@ -52,17 +52,17 @@ size_t Metadata::GetBatchCnt() const {
 }
 
 
-void Metadata::SetNowOffset(int64_t offset) {
+void Metadata::SetNowOffset(uint64_t offset) {
     now_offset_ = offset;
 }
 
 std::vector<PhysTypeVariant> Metadata::Serialize() const {
-    int64_t old_offset = now_offset_;
+    uint64_t old_offset = now_offset_;
     std::vector<PhysTypeVariant> result = schema_.Serialize();
 
     result.reserve(result.size() + columns_cnt_.size() * 3 + 2);
 
-    result.emplace_back(static_cast<int64_t>(columns_cnt_.size()));
+    result.emplace_back(columns_cnt_.size());
 
     for (size_t i = 0; i < columns_cnt_.size(); i++) {
         result.emplace_back(batch_offsets_[i]);
